Actualizar informacion de una fuente raw

Actualiza 'fecha_descarga' y 'fecha_actualizar' de una fuente en la sheet de fuentes en el drive de Argendata. Hace drive_upload() con overwrite = T pisando la version anterior de la fuente en el drive.

actualizar_fuente_raw(id_fuente, fecha_actualizar = NULL)

Arguments

id_fuente

integer id numerico que permite seleccionar la fuente segun aparece en el sheet. Para consultar ids usar fuentes_raw()

fecha_actualizar

string or date Valor de la fecha en que se debe actualizar la descarga. Puede ser clase 'date' como producto de Sys.Date() o un string parseable por as.Date()

Details

La fecha será actualizada usando Sys.time() al momento de su ejecución.