Agregar nueva fuente clean

Agrega una fuente no registrada previamente: genera una nueva entrada en la sheet de fuentes y hace drive_upload() con overwrite = F de la fuente.

agregar_fuente_clean(
  id_fuente_raw = NULL,
  path_clean = NULL,
  nombre = NULL,
  script = NULL,
  prompt = TRUE
)

Arguments

id_fuente_raw

integer id numerico que permite seleccionar la fuente raw segun aparece en el sheet. Para consultar ids usar fuentes_raw()

path_clean

string Nombre del archivo de la fuente tal cual fue guardado en el directorio data/_FUENTES/clean/ de argendata-etl

nombre

string Nombre único que identifica a la fuente en su versión 'clean'.

script

string Nombre del archivo del script de descarga de la fuente tal cual se guardó en scripts/limpieza_fuentes/ de argendata-etl

prompt

logical Si es TRUE (default) evalua si ya fuentes clean referidas al id_fuente_raw y pide confirmacoion antes de continuar.

Details

Más de una fuente clean puede referir a una misma fuente raw. Por ejemplo, si la fuente raw consiste en un excel de multiples hojas, cada hoja debería pasar a ser un csv independiente.